requirements.txt
flask>=3.0,<4.0
Flask-CORS>=4.0,<5.0
requests>=2.31,<3.0
Pillow>=9.5,<10.0
exifread>=2.3,<3.0
watchdog>=2.3,<3.0
APScheduler>=3.10,<4.0
pyyaml>=6.0,<7.0
openai>=1.0,<2.0
pymilvus>=2.4,<3.0
torch>=2.0,<3.0
